backend/pyproject.toml
[project]
name = "roleready-backend"
version = "0.1.0"
description = "FastAPI backend with PydanticAI agent for RoleReady"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-ai>=0.0.13",
    "python-multipart>=0.0.9",
    "pypdf>=4.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.uvicorn]
factory = true
workers = 1


